+++ ChangeLog.md
@@ -12,6 +12,7 @@
     - Change TTP::commandExec() prototype to more standard '$command, $opts' (todo #86)
     - Deduplicate TTP::runnerDaemon::_mqtt_timeout() and TTP::runnerDaemon::messagingTimeout() (todo #83)
     - Bump minor candidate version number due to site.json change
+    - RunnerDaemon better check listeningPort (todo #81) and listeningInterval (todo #82)
 
 ### 4.4.1